Support for long-vector columns (fixes #201)

+++ src/fst_column_factory.h
@@ -43,37 +43,37 @@ class ColumnFactory : public IColumnFactory
 public:
   ~ColumnFactory() {};
 
-  IFactorColumn* CreateFactorColumn(int nrOfRows, int nrOfLevels, FstColumnAttribute columnAttribute)
+  IFactorColumn* CreateFactorColumn(uint64_t nrOfRows, uint64_t nrOfLevels, FstColumnAttribute columnAttribute)
   {
     return new FactorColumn(nrOfRows, nrOfLevels, columnAttribute);
   }
 
-  ILogicalColumn* CreateLogicalColumn(int nrOfRows, FstColumnAttribute columnAttribute)
+  ILogicalColumn* CreateLogicalColumn(uint64_t nrOfRows, FstColumnAttribute columnAttribute)
   {
     return new LogicalColumn(nrOfRows);
   }
 
-  IDoubleColumn* CreateDoubleColumn(int nrOfRows, FstColumnAttribute columnAttribute, short int scale)
+  IDoubleColumn* CreateDoubleColumn(uint64_t nrOfRows, FstColumnAttribute columnAttribute, short int scale)
   {
     return new DoubleColumn(nrOfRows, columnAttribute, scale);
   }
 
-  IIntegerColumn* CreateIntegerColumn(int nrOfRows, FstColumnAttribute columnAttribute, short int scale)
+  IIntegerColumn* CreateIntegerColumn(uint64_t nrOfRows, FstColumnAttribute columnAttribute, short int scale)
   {
     return new IntegerColumn(nrOfRows, columnAttribute, scale);
   }
 
-  IByteColumn* CreateByteColumn(int nrOfRows, FstColumnAttribute columnAttribute)
+  IByteColumn* CreateByteColumn(uint64_t nrOfRows, FstColumnAttribute columnAttribute)
   {
     return new ByteColumn(nrOfRows);
   }
 
-  IInt64Column* CreateInt64Column(int nrOfRows, FstColumnAttribute columnAttribute, short int scale)
+  IInt64Column* CreateInt64Column(uint64_t nrOfRows, FstColumnAttribute columnAttribute, short int scale)
   {
     return new Int64Column(nrOfRows, columnAttribute, scale);
   }
 
-  IStringColumn* CreateStringColumn(int nrOfRows, FstColumnAttribute columnAttribute)
+  IStringColumn* CreateStringColumn(uint64_t nrOfRows, FstColumnAttribute columnAttribute)
   {
     return new BlockReaderChar();
   }
